+++ runtime/src/network/server.hpp
@@ -6,7 +6,7 @@
 #include "network/connection_state.hpp"
 #include "network/net_common.hpp"
 #include "network/packet.hpp"
-#include "shared/scene/cnet_component.hpp"
+#include "shared/scene/component/cnet_component.hpp"
 #include "shared/types.hpp"
 #include <optional>
 #include <steam/isteamnetworkingutils.h>
