+++ runtime/src/world.hpp
@@ -5,9 +5,9 @@
 #include "network/server.hpp"
 #include "network/util.hpp"
 #include "shared/app.hpp"
-#include "shared/scene/cnet_component.hpp"
+#include "shared/scene/component/cnet_component.hpp"
 #include "shared/scene/component/crotor.hpp"
-#include "shared/scene/fps_camera.hpp"
+#include "shared/scene/component/fps_camera.hpp"
 #include "shared/state/player_input.hpp"
 #include "shared/utility/unique_id.hpp"
 #include <GUI/BsGUILabel.h>
