+++ shared/src/shared/scene/builder.cpp
@@ -29,13 +29,13 @@
 #include "shared/asset.hpp"
 #include "shared/log.hpp"
 #include "shared/render/shader.hpp"
-#include "shared/scene/cnet_component.hpp"
+#include "shared/scene/component/cnet_component.hpp"
 #include "shared/scene/component/cspline.hpp"
 #include "shared/scene/component/ctag.hpp"
 #include "shared/scene/component/cwind.hpp"
 #include "shared/scene/component/cwind_affectable.hpp"
 #include "shared/scene/component/crotor.hpp"
-#include "shared/scene/fps_walker.hpp"
+#include "shared/scene/component/fps_walker.hpp"
 #include "shared/state/moveable_state.hpp"
 #include "shared/utility/util.hpp"
 #include "shared/wind/base_functions.hpp"
