+++ shared/src/shared/scene/component/cnet_component.hpp
